update to v2.3.0 schema

+++ models/nested_frozen_disks.go
@@ -21,7 +21,7 @@ type NestedFrozenDisks struct {
 
 	// boot
 	// Required: true
-	Boot *int64 `json:"boot"`
+	Boot *int32 `json:"boot"`
 
 	// bus
 	// Required: true
